requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.2
cyclopts==3.24.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.119.0
fastmcp==2.12.4
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
isodate==0.7.2
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.17.0
mdurl==0.1.2
memvid==0.1.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
openai==2.3.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
parse==1.20.2
pathable==0.4.4
pillow==12.0.0
pycparser==2.23
pydantic==2.12.2
pydantic-core==2.41.4
pydantic-settings==2.11.0
pygments==2.19.2
pypdf2==3.0.1
pyperclip==1.11.0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
qrcode==8.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
werkzeug==3.1.1
